Update mavenCentral_cd.yml (#1154)
* Update mavenCentral_cd.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* Update codeql-analysis.yml


* Update mavenCentral_cd.yml


* Update pom.xml

https://github.com/orgs/community/discussions/27165?sort=top


* Update mavenCentral_cd.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* Update mavenCentral_cd.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* Update codeql-analysis.yml


* Update mavenCentral_cd.yml


* Update pom.xml

https://github.com/orgs/community/discussions/27165?sort=top


* Update mavenCentral_cd.yml


* Update mavenCentral_cd.yml


* Update codeql-analysis.yml


* rebase onto master and merge


* adding more environment variables


* gpg: invalid pinentry mode '--quiet'


* gpg: signing failed: "No pinentry"


* Invalid workflow file


* gpg: signing failed: Inappropriate ioctl for device


* gpg: signing failed


* brew: command not found


* . ~/.bash_profile


* . ~/.bash_profile


* disable maven caching


* remove failing signing test


* --pinentry-mode loopback


* -DskipTests


* version bump 20230609 for e2e release test


* fix cleanSummaryReportsDirectoryBeforeExecution


---------

+++ src/main/java/com/shaft/tools/io/internal/ReportManagerHelper.java
@@ -953,8 +953,10 @@ public class ReportManagerHelper {
     }
 
     public static void cleanExecutionSummaryReportDirectory() {
-        String executionSummaryReportFolderPath = SHAFT.Properties.paths.executionSummaryReport();
-        FileActions.getInstance().deleteFolder(executionSummaryReportFolderPath.substring(0, executionSummaryReportFolderPath.length() - 1));
+        if (SHAFT.Properties.reporting.cleanSummaryReportsDirectoryBeforeExecution()) {
+            String executionSummaryReportFolderPath = SHAFT.Properties.paths.executionSummaryReport();
+            FileActions.getInstance().deleteFolder(executionSummaryReportFolderPath.substring(0, executionSummaryReportFolderPath.length() - 1));
+        }
     }
 
     public static void openExecutionSummaryReportAfterExecution() {
